+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ## Added
 - Added condition `has_obtained_quill` with 3 simple goals
+- Added condition `has_obtained_map` with 28 simple goals
 
 </details>